рефакторинг App, MenuBar, AddMeme, Meme,  удален компонент Main, вынес в отдельный компонент Copyright, routs
