+++ client/src/components/App/AppTemplate.tsx
@@ -1,25 +1,29 @@
 import React from 'react'
+import { Provider } from 'react-redux'
 import { BrowserRouter as Router } from 'react-router-dom'
-import { useRoutes } from '../../routes'
-import { MenuAppBar } from './../MenuAppBar'
 import { useSelector } from 'react-redux'
+
+import { configureStore } from '../../redux/authToolkitRedux'
+import { useRoutes } from '../../customHooks/routes'
+import { MenuBar } from '../MenuBar'
 import { RootState } from '../../redux/authToolkitRedux/StoreSlices'
-import './App.scss'
+import '../../styles/App.scss'
 import 'materialize-css'
 
-function App() {
+const store = configureStore()
 
+export function AppTemplate() {
   // const isAuthenticated = useSelector((state: RootState) => state.authorization.isAuthenticated)
   //TODO для разработки, убрать
   const isAuthenticated = true
   const routes = useRoutes(isAuthenticated)
 
   return (
+    <Provider store={store}>
       <Router>
-        {isAuthenticated && <MenuAppBar/>}
+        {isAuthenticated && <MenuBar/>}
         <div className="app">{routes}</div>
       </Router>
+    </Provider>
   )
 }
-
-export default App
