+++ client/src/components/Copyright/index.ts
@@ -1,0 +1,1 @@
+export { CopyrightTemplate as Copyright } from './CopyrightTemplate'
