+++ client/src/components/MenuBar/index.ts
@@ -1,0 +1,1 @@
+export { MenuBarBehavior as MenuBar } from './MenuBarBehavior'
