+++ client/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import { AppTemp } from './components/App/AppTemp'
-import './index.css'
+import { AppTemplate } from './components/App/AppTemplate'
+import './styles/index.css'
 
-ReactDOM.render(<AppTemp />, document.getElementById('root'))
+ReactDOM.render(<AppTemplate />, document.getElementById('root'))
