+++ client/src/pages/AddMeme/index.ts
@@ -1,0 +1,1 @@
+export { AddMemeBehavior as AddMeme } from './AddMemeBehavior'
