+++ client/src/pages/Meme/index.ts
