+++ client/src/redux/authToolkitRedux/StoreSlices/authorization.ts
@@ -1,4 +1,5 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
+
 import { authLoginFetch } from '../../../API/authApi'
 
 export interface Credentials {
