+++ client/src/styles/memeStyle.ts
