+++ client/src/styles/menuBarStyle.ts
@@ -1,0 +1,15 @@
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
+
+export const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      flexGrow: 1,
+    },
+    menuButton: {
+      marginRight: theme.spacing(2),
+    },
+    title: {
+      flexGrow: 1,
+    },
+  })
+)
